+++ packages/gatsby-tinacms-remark/gatsby-node.js
@@ -20,7 +20,10 @@ const slash = require("slash");
 
 exports.setFieldsOnGraphQLNodeType = ({ type }) => {
   const pathRoot = slash(process.cwd());
-  if (type.name === `MarkdownRemark`) {
+
+  const hasMarkdown = !!type.nodes.find(node => node.internal.owner === 'gatsby-transformer-remark')
+
+  if (hasMarkdown) {
     return {
       rawFrontmatter: {
         type: GraphQLString,