+++ packages/tinacms/src/components/sidebar/Sidebar.tsx
@@ -22,13 +22,15 @@ import { useState } from 'react'
 import { StyledFrame } from '../SyledFrame'
 import styled, { keyframes, css } from 'styled-components'
 import { FormsView } from '../FormView'
-import { Modal } from '../modals/ModalProvider'
+import { Modal, ModalHeader, ModalBody } from '../modals/ModalProvider'
 import { ModalFullscreen } from '../modals/ModalFullscreen'
+import { ModalPopup } from '../modals/ModalPopup'
 import {
   HamburgerIcon,
   LeftArrowIcon,
   EditIcon,
   TinaIcon,
+  CloseIcon,
 } from '@tinacms/icons'
 import {
   GlobalStyles,
@@ -107,17 +109,60 @@ export const Sidebar = () => {
             <Watermark />
           </MenuPanel>
           {ActiveView && (
-            <Modal>
-              <ModalFullscreen>
-                <button onClick={() => setActiveView(null)}>Close Modal</button>
-                <ActiveView.Component />
-              </ModalFullscreen>
-            </Modal>
+            <ActiveViewModal
+              name={ActiveView.name}
+              close={() => setActiveView(null)}
+              layout={ActiveView.layout}
+            >
+              <ActiveView.Component />
+            </ActiveViewModal>
           )}
         </SidebarWrapper>
       </StyledFrame>
       <SidebarToggle {...sidebar} />
     </SidebarContainer>
+  )
+}
+
+interface ActiveViewProps {
+  children: any
+  name: string
+  close: any
+  layout?: 'fullscreen' | 'popup'
+}
+
+const ActiveViewModal = ({
+  children,
+  name,
+  close,
+  layout,
+}: ActiveViewProps) => {
+  let Wrapper
+
+  switch (layout) {
+    case 'popup':
+      Wrapper = ModalPopup
+      break
+    case 'fullscreen':
+      Wrapper = ModalFullscreen
+      break
+    default:
+      Wrapper = ModalPopup
+      break
+  }
+
+  return (
+    <Modal>
+      <Wrapper>
+        <ModalHeader>
+          {name}
+          <CloseButton onClick={close}>
+            <CloseIcon />
+          </CloseButton>
+        </ModalHeader>
+        <ModalBody>{children}</ModalBody>
+      </Wrapper>
+    </Modal>
   )
 }
 
@@ -426,4 +471,19 @@ const SidebarContainer = styled.div<{ open: boolean }>`
     0
   ) !important;
   pointer-events: none;
+`
+
+const CloseButton = styled.div`
+  display: flex;
+  align-items: center;
+  fill: ${color.grey(3)};
+  cursor: pointer;
+  transition: fill 85ms ease-out;
+  svg {
+    width: 1.5rem;
+    height: auto;
+  }
+  &:hover {
+    fill: ${color.grey(8)};
+  }
 `