+++ packages/demo-cra/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.7-alpha.1...demo-cra@1.0.7) (2019-11-18)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.7-alpha.1](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.6...demo-cra@1.0.7-alpha.1) (2019-11-18)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.7-alpha.0](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.6...demo-cra@1.0.7-alpha.0) (2019-11-18)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
 ## [1.0.6](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.5...demo-cra@1.0.6) (2019-11-14)
 
 **Note:** Version bump only for package demo-cra